==> labs/interpolation/04-nn-and-idw.Rmd <==
# Near neighbour and inverse-distance weighted interpolation
Run this first to make sure all the data and packages you need are loaded. If any data are missing you probably didn't make them in one of the previous instruction pages.

```{r}
library(sf)
library(tmap)
library(terra)
library(dplyr)
library(gstat)

volcano <- rast("data/maungawhau.tif")
names(volcano) <- "data/height"

controls <- st_read("data/controls.gpkg")
sites_sf <- st_read("data/sites-sf.gpkg")
sites_raster <- rast("data/sites-raster.tif")
```

## Inverse-distance weighted interpolation
These two methods are very similar, and IDW is actually *more general* so we'll show it first.

As with all the `gstat` methods we use the `gstat::gstat` function to make a statistical model, and then apply it using the `raster::interpolate` function.

```{r}
fit_IDW <- gstat(                   # makes a model 
  formula = height ~ 1,             # The column `height` is what we are interested in
  data = as(controls, "Spatial"),   # using sf but converting to sp, which is required
  set = list(idp = 2),
  # nmax = 12, maxdist = 100        # you can experiment with these options later...
)
```

The `idp` setting here is the inverse-distance power used in the calculation. Once you understand what is going on in general, you should experiment with this, and also with `nmax` (the maximum number of control points to include in any estimate) and `maxdist` (the maximum distance to any control point to use in an estimate) to see how the results change.

Having made the model (called `fit_IDW`) we pass it to the `predict` function to obtain interpolated values (called `var1.pred`) at the locations specified by `sites` and then finally convert this to a raster for visualization.

```{r}
interp_pts_IDW <- predict(fit_IDW, sites_sf)
interp_IDW <- rasterize(as(interp_pts_IDW, "SpatVector"), sites_raster, "var1.pred")
names(interp_IDW) <- "height" # rename the variable to something more friendly
```

And then we can view the outcome in the usual ways. 
```{r}
interp_IDW
```

And we can map the result

```{r}
tm_shape(interp_IDW) + 
  tm_raster(pal = "-BrBG", style = "cont") +
  tm_legend(outside = TRUE)
```

```{r}
persp(interp_IDW, scale = FALSE, expand = 2, theta = 35, phi = 30, lwd = 0.5)
```

### Nearest neighbour
The basic model above can be parameterised differently, setting `nmax` to 1 to make a simple nearest neighbour (i.e. proximity polygon) interpolation:

```{r}
fit_NN <- gstat( # makes a model 
  formula = height ~ 1,
  data = as(controls, "Spatial"), 
  nmax = 1, # by setting nmax to 1 we force it to 1, and get nearest neighbour
)

# and interpolate like before
interp_pts_NN <- predict(fit_NN, sites_sf)
interp_NN <- rasterize(as(interp_pts_NN, "SpatVector"), sites_raster, "var1.pred")
names(interp_NN) <- "height"

# and display it
persp(interp_NN, scale = FALSE, expand = 2, theta = 35, phi = 30, lwd = 0.5)
```

We can confirm this matches with proximity polygons like this:
```{r}
tm_shape(interp_NN) + 
  tm_raster(pal = "-BrBG", style = "cont") + 
  tm_shape(st_voronoi(st_union(controls))) + 
  tm_borders(col = "yellow", lwd = 0.5) + 
  tm_shape(controls) + 
  tm_dots() + 
  tm_legend(outside = TRUE)
```

Back to [data prep](03-preparing-for-interpolation.md) | On to [trend surfaces and kriging](05-trend-surfaces-and-kriging.md)
